+++ src/tinygp/kernels/base.py
@@ -114,7 +114,7 @@ class Kernel(metaclass=ABCMeta):
             return Sum(self, other)
         return Sum(self, Constant(other))
 
-    def __radd__(self, other: Union["Kernel", JAXArray]) -> "Kernel":
+    def __radd__(self, other: Any) -> "Kernel":
         # We'll hit this first branch when using the `sum` function
         if other == 0:
             return self
@@ -127,7 +127,7 @@ class Kernel(metaclass=ABCMeta):
             return Product(self, other)
         return Product(self, Constant(other))
 
-    def __rmul__(self, other: Union["Kernel", JAXArray]) -> "Kernel":
+    def __rmul__(self, other: Any) -> "Kernel":
         if isinstance(other, Kernel):
             return Product(other, self)
         return Product(Constant(other), self)
@@ -173,7 +173,7 @@ class Custom(Kernel):
     function: Callable[[Any, Any], Any]
 
     def evaluate(self, X1: JAXArray, X2: JAXArray) -> JAXArray:
-        return self.function(X1, X2)  # type: ignore
+        return self.function(X1, X2)
 
 
 @dataclass
@@ -214,11 +214,9 @@ class Constant(Kernel):
 
     value: JAXArray
 
-    def __post_init__(self) -> None:
+    def evaluate(self, X1: JAXArray, X2: JAXArray) -> JAXArray:
         if jnp.ndim(self.value) != 0:
             raise ValueError("The value of a constant kernel must be a scalar")
-
-    def evaluate(self, X1: JAXArray, X2: JAXArray) -> JAXArray:
         return self.value
 
 
@@ -234,6 +232,8 @@ class DotProduct(Kernel):
     """
 
     def evaluate(self, X1: JAXArray, X2: JAXArray) -> JAXArray:
+        if jnp.ndim(X1) == 0:
+            return X1 * X2
         return X1 @ X2