+++ tests/test_kernels/test_kernels.py
@@ -25,13 +25,16 @@ def data(random):
 def test_constant(data):
     x1, x2 = data
 
-    # Check for dimension issues, either when instantiated...
+    # Check for dimension issues when evaluated
     with pytest.raises(ValueError):
-        kernels.Constant(jnp.ones(3))
+        k1 = kernels.Constant(jnp.ones(3))
+        v = jnp.ones(3)
+        k1.evaluate(v, v)
 
-    # ... or when multiplied
+    # Check for dimension issues when multiplied and evaluated.
+    k = jnp.ones(3) * kernels.Matern32(1.5)
     with pytest.raises(ValueError):
-        jnp.ones(3) * kernels.Matern32(1.5)
+        k.evaluate(v, v)
 
     # Check that multiplication has the expected behavior
     factor = 2.5
@@ -82,3 +85,12 @@ def test_conditioned(data):
             cond(x1, x2),
             k2(x1, x2) - k2(x1, x1) @ jnp.linalg.solve(K, k2(x1, x2)),
         )
+
+
+def test_dot_product(data):
+    x1, x2 = data
+    kernel = kernels.DotProduct()
+    np.testing.assert_allclose(kernel(x1, x2), jnp.dot(x1, x2.T))
+    np.testing.assert_allclose(
+        kernel(x1[:, 0], x2[:, 0]), x1[:, 0][:, None] * x2[:, 0][None]
+    )